api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.4
yfinance==0.2.40
slowapi==0.1.9
prometheus-client==0.20.0
websockets==12.0
azure-servicebus
python-dotenv
sentry-sdk[fastapi]
